+++ libs/expo/betterangels/src/lib/hooks/user/useSignOut.ts
@@ -1,5 +1,5 @@
 import { gql, useMutation } from '@apollo/client';
-import { useApiConfig } from '@monorepo/expo/shared/providers';
+import { useApiConfig } from '@monorepo/expo/shared/apollo';
 import { useCallback } from 'react';
 import useUser from './useUser';
