+++ libs/expo/shared/clients/jest.config.ts
@@ -7,5 +7,5 @@ export default {
     '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../../coverage/libs/expo/shared/apollo',
+  coverageDirectory: '../../../../coverage/libs/expo/shared/clients',
 };
