+++ libs/expo/shared/clients/project.json
@@ -1,7 +1,7 @@
 {
   "name": "expo-shared-apollo",
   "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/expo/shared/apollo/src",
+  "sourceRoot": "libs/expo/shared/clients/src",
   "projectType": "library",
   "tags": [],
   "targets": {
@@ -9,7 +9,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/expo/shared/apollo/jest.config.ts"
+        "jestConfig": "libs/expo/shared/clients/jest.config.ts"
       }
     }
   }
