+++ libs/expo/shared/clients/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './apollo';
+export * from './http';
