--- libs/expo/shared/providers/README.md
@@ -1,11 +1,0 @@
-# expo-shared-providers
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build expo-shared-providers` to build the library.
-
-## Running unit tests
-
-Run `nx test expo-shared-providers` to execute the unit tests via [Jest](https://jestjs.io).
