--- libs/expo/shared/providers/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@monorepo/expo/shared/providers",
-  "version": "0.0.1",
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "type": "commonjs",
-  "main": "./src/index.js",
-  "typings": "./src/index.d.ts",
-  "private": true
-}
