--- libs/expo/shared/providers/project.json
@@ -1,26 +1,0 @@
-{
-  "name": "expo-shared-providers",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/expo/shared/providers/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/expo/shared/providers",
-        "main": "libs/expo/shared/providers/src/index.ts",
-        "tsConfig": "libs/expo/shared/providers/tsconfig.lib.json",
-        "assets": ["libs/expo/shared/providers/*.md"]
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/expo/shared/providers/jest.config.ts"
-      }
-    }
-  }
-}
