--- libs/expo/shared/providers/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/ApiConfigProvider';
