--- libs/expo/shared/providers/tsconfig.lib.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../../../dist/out-tsc",
-    "declaration": true,
-    "types": ["node"]
-  },
-  "include": [
-    "src/**/*.ts",
-    "src/lib/expo-shared-providers.spec.tsx",
-    "src/lib/ApiConfigProvider.tsx"
-  ],
-  "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
-}
