+++ tsconfig.base.json
@@ -18,11 +18,8 @@
     "paths": {
       "@betterangels/*": ["apps/betterangels/*"],
       "@monorepo/expo/betterangels": ["libs/expo/betterangels/src/index.ts"],
-      "@monorepo/expo/shared/apollo": ["libs/expo/shared/apollo/src/index.ts"],
+      "@monorepo/expo/shared/apollo": ["libs/expo/shared/clients/src/index.ts"],
       "@monorepo/expo/shared/icons": ["libs/expo/shared/icons/src/index.ts"],
-      "@monorepo/expo/shared/providers": [
-        "libs/expo/shared/providers/src/index.ts"
-      ],
       "@monorepo/expo/shared/static": ["libs/expo/shared/static/src/index.ts"],
       "@monorepo/expo/shared/ui-components": [
         "libs/expo/shared/ui-components/src/index.ts"
